validacion completa de eliminar y actualizar de todos los EndPoint

+++ src/main/java/com/example/seguimientocuerpos/controller/ActualizacionPEController.java
@@ -3,6 +3,8 @@ package com.example.seguimientocuerpos.controller;
 import com.example.seguimientocuerpos.model.ActualizacionPE;
 import com.example.seguimientocuerpos.service.ActualizacionPEService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -29,13 +31,19 @@ public class ActualizacionPEController {
         return actualizacionPEService.findById(id).get();
     }
 
-    @DeleteMapping("/actualizacionPE/{id}")
-    public void deleteById(@PathVariable String id){
-        actualizacionPEService.deleteById(id);
+    @DeleteMapping("/actualizacionPE/{id}/{id_usuario}")
+    public ResponseEntity<Object> deleteById(@PathVariable String id, @PathVariable String id_usuario){
+        if(actualizacionPEService.findById(id).get().getId_usuario().equals(id_usuario)){
+            actualizacionPEService.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
-    @PutMapping("/actualizacionPE")
-    public void update(@RequestBody ActualizacionPE actualizacionPE){
-        actualizacionPEService.save(actualizacionPE);
+    @PutMapping("/actualizacionPE/{id_usuario}")
+    public ResponseEntity<Object> update(@RequestBody ActualizacionPE actualizacionPE, @PathVariable String id_usuario){
+        if(actualizacionPEService.findById(actualizacionPE.getId()).get().getId_usuario().equals(id_usuario)){
+            actualizacionPEService.save(actualizacionPE);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 }
