+++ src/main/java/com/example/seguimientocuerpos/controller/ApoyoEconomicoController.java
@@ -3,6 +3,8 @@ package com.example.seguimientocuerpos.controller;
 import com.example.seguimientocuerpos.model.ApoyoEconomico;
 import com.example.seguimientocuerpos.service.ApoyoEconomicoService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -29,13 +31,19 @@ public class ApoyoEconomicoController {
         return apoyoEconomicoService.findById(id).get();
     }
 
-    @DeleteMapping("/apoyoEconomico/{id}")
-    public void deleteById(@PathVariable String id){
-        apoyoEconomicoService.deleteById(id);
+    @DeleteMapping("/apoyoEconomico/{id}/{id_usuario}")
+    public ResponseEntity<Object> deleteById(@PathVariable String id, @PathVariable String id_usuario){
+        if(apoyoEconomicoService.findById(id).get().getId_usuario().equals(id_usuario)){
+            apoyoEconomicoService.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
-    @PutMapping("/apoyoEconomico")
-    public void update(@RequestBody ApoyoEconomico apoyoEconomico){
-        apoyoEconomicoService.save(apoyoEconomico);
+    @PutMapping("/apoyoEconomico/{id_usuario}")
+    public ResponseEntity<Object> update(@RequestBody ApoyoEconomico apoyoEconomico, @PathVariable String id_usuario){
+        if(apoyoEconomicoService.findById(apoyoEconomico.getId()).get().getId_usuario().equals(id_usuario)){
+            apoyoEconomicoService.save(apoyoEconomico);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 }
