+++ src/main/java/com/example/seguimientocuerpos/controller/ColaboracionCAController.java
@@ -3,6 +3,8 @@ package com.example.seguimientocuerpos.controller;
 import com.example.seguimientocuerpos.model.ColaboracionCA;
 import com.example.seguimientocuerpos.service.ColaboracionCAService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -29,13 +31,19 @@ public class ColaboracionCAController {
         return colaboracionCAService.findById(id).get();
     }
 
-    @DeleteMapping("/colaboracionCA/{id}")
-    public void deleteById(@PathVariable String id){
-        colaboracionCAService.deleteById(id);
+    @DeleteMapping("/colaboracionCA/{id}/{id_usuario}")
+    public ResponseEntity<Object> deleteById(@PathVariable String id, @PathVariable String id_usuario){
+        if(colaboracionCAService.findById(id).get().getId_usuario().equals(id_usuario)){
+            colaboracionCAService.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
-    @PutMapping("/colaboracionCA")
-    public void update(@RequestBody ColaboracionCA colaboracionCA){
-        colaboracionCAService.save(colaboracionCA);
+    @PutMapping("/colaboracionCA/{id_usuario}")
+    public ResponseEntity<Object> update(@RequestBody ColaboracionCA colaboracionCA, @PathVariable String id_usuario) {
+        if (colaboracionCAService.findById(colaboracionCA.getId()).get().getId_usuario().equals(id_usuario)) {
+            colaboracionCAService.save(colaboracionCA);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 }
