+++ src/main/java/com/example/seguimientocuerpos/controller/DireccionIndividualController.java
@@ -3,6 +3,8 @@ package com.example.seguimientocuerpos.controller;
 import com.example.seguimientocuerpos.model.DireccionIndividual;
 import com.example.seguimientocuerpos.service.DireccionIndividualService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -29,13 +31,19 @@ public class DireccionIndividualController {
         return direccionIndividualService.findById(id).get();
     }
 
-    @DeleteMapping("/direccionIndividual")
-    public void deleteById(@PathVariable String id){
-        direccionIndividualService.deleteById(id);
+    @DeleteMapping("/direccionIndividual/{id}/{id_usuario}")
+    public ResponseEntity<Object> deleteById(@PathVariable String id, @PathVariable String id_usuario){
+        if(direccionIndividualService.findById(id).get().getId_usuario().equals(id_usuario)){
+            direccionIndividualService.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
-    @PutMapping("/direccionIndividual")
-    public void update(@RequestBody DireccionIndividual direccionIndividual){
-        direccionIndividualService.save(direccionIndividual);
+    @PutMapping("/direccionIndividual/{id_usuario}")
+    public ResponseEntity<Object> update(@RequestBody DireccionIndividual direccionIndividual, @PathVariable String id_usuario){
+        if(direccionIndividualService.findById(direccionIndividual.getId()).get().getId_usuario().equals(id_usuario)){
+            direccionIndividualService.save(direccionIndividual);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 }
