+++ src/main/java/com/example/seguimientocuerpos/controller/ProyectosColectivosController.java
@@ -3,6 +3,8 @@ package com.example.seguimientocuerpos.controller;
 import com.example.seguimientocuerpos.model.ProyectosColectivos;
 import com.example.seguimientocuerpos.service.ProyectosColectivosService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -29,13 +31,19 @@ public class ProyectosColectivosController {
         return proyectosColectivosService.findById(id).get();
     }
 
-    @DeleteMapping("/proyectoColectivo/{id}")
-    public void deleteById(@PathVariable String id){
-        proyectosColectivosService.deleteById(id);
+    @DeleteMapping("/proyectoColectivo/{id}/{id_usuario}")
+    public ResponseEntity<Object> deleteById(@PathVariable String id, @PathVariable String id_usuario){
+        if(proyectosColectivosService.findById(id).get().getId_usuario().equals(id_usuario)){
+            proyectosColectivosService.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
-    @PutMapping("/proyectoColectivo")
-    public void update(@RequestBody ProyectosColectivos proyectosColectivos){
-        proyectosColectivosService.save(proyectosColectivos);
+    @PutMapping("/proyectoColectivo/{id_usuario}")
+    public ResponseEntity<Object> update(@RequestBody ProyectosColectivos proyectosColectivos, @PathVariable String id_usuario){
+        if(proyectosColectivosService.findById(proyectosColectivos.getId()).get().getId_usuario().equals(id_usuario)){
+            proyectosColectivosService.save(proyectosColectivos);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 }
